requirements.txt
flask==2.1.0
gunicorn==20.0.4
python-telegram-bot==13.3
telegram-menu==0.3.4
google_auth_oauthlib
google-api-python-client